+++ src/check_jsonschema/cli/main_command.py
@@ -74,10 +74,10 @@ including the following formats by default:
     date, date-time, email, ipv4, ipv6, regex, uuid
 
 \b
-For the "regex" format, there are multiple modes which can be specified with
-'--format-regex':
-    default  |  check that the string is a valid ECMAScript regex
-    python   |  check that the string is a valid python regex
+For handling of regexes, there are multiple modes which can be specified with
+'--regex-variant':
+    default  |  use ECMAScript regex syntax (via regress)
+    python   |  use python regex syntax
 
 \b
 The '--builtin-schema' flag supports the following schema names:
@@ -148,11 +148,18 @@ The '--disable-formats' flag supports the following formats:
 )
 @click.option(
     "--format-regex",
+    hidden=True,
+    help="Legacy name for `--regex-variant`.",
+    default=None,
+    type=click.Choice([x.value for x in RegexVariantName], case_sensitive=False),
+)
+@click.option(
+    "--regex-variant",
     help=(
-        "Set the mode of format validation for regexes. "
-        "If `--disable-formats regex` is used, this option has no effect."
+        "Name of which regex dialect should be used for format checking "
+        "and 'pattern' matching."
     ),
-    default=RegexVariantName.default.value,
+    default=None,
     type=click.Choice([x.value for x in RegexVariantName], case_sensitive=False),
 )
 @click.option(
@@ -240,7 +247,8 @@ def main(
     no_cache: bool,
     cache_filename: str | None,
     disable_formats: tuple[list[str], ...],
-    format_regex: Literal["python", "default"],
+    format_regex: Literal["python", "default"] | None,
+    regex_variant: Literal["python", "default"] | None,
     default_filetype: Literal["json", "yaml", "toml", "json5"],
     traceback_mode: Literal["full", "short"],
     data_transform: Literal["azure-pipelines", "gitlab-ci"] | None,
@@ -252,6 +260,8 @@ def main(
     instancefiles: tuple[t.IO[bytes], ...],
 ) -> None:
     args = ParseResult()
+
+    args.set_regex_variant(regex_variant, legacy_opt=format_regex)
 
     args.set_schema(schemafile, builtin_schema, check_metaschema)
     args.set_validator(validator_class)
@@ -267,7 +277,6 @@ def main(
     else:
         args.disable_formats = normalized_disable_formats
 
-    args.format_regex = RegexVariantName(format_regex)
     args.disable_cache = no_cache
     args.default_filetype = default_filetype
     args.fill_defaults = fill_defaults